ソースコードを大々的に書き直し
