+++ index.html
@@ -25,7 +25,7 @@
     init()
   </script>
   <body>
-    [↑][↓][←][→]：移動<br>
-    ただひたすら迷路を踏破するだけのもの。<br>
+    [⇦][⇩][⇦][⇨]：移動、[Esc]：一時停止<br>
+    ただひたすら上端でくるくるしてる◆を目指して迷路を踏破するだけのもの。<br>
   </body>    
 </html>
